v2_chain_google_docling/requirements.txt
python-dotenv
docling
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
chromadb
fastapi 
uvicorn

